requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.0.0
django-filter==23.2
djangorestframework-simplejwt==5.2.2

# Database
psycopg2cffi==2.9.0
dj-database-url==2.1.0

# Task Queue & Background Jobs
celery==5.3.1
redis==4.6.0
django-celery-beat==2.5.0
django-celery-results==2.6.0

# WebSockets for real-time features
channels==4.0.0
channels-redis==4.1.0

# Authentication & Security
django-allauth==0.57.0
django-oauth-toolkit==2.3.0
PyJWT==2.8.0
bcrypt==4.0.1
cryptography==41.0.4

# File Storage
boto3==1.28.57
django-storages==1.14.2

# API Documentation
drf-spectacular==0.26.4
drf-yasg==1.21.7

# Development & Testing
pytest==7.4.2
pytest-django==4.5.2
factory-boy==3.2.1
faker==19.6.2
coverage==7.3.1
black==23.9.1
isort==5.12.0
flake8==6.0.0
mypy==1.5.1

# Frontend build (for PWA)
django-webpack-loader==2.0.1

# AI/ML (placeholder for future integration)
# scikit-learn==1.3.1  # Temporarily disabled due to Python 3.13 compatibility
# pandas==2.1.1
# numpy==1.25.2

# Monitoring & Observability
sentry-sdk==1.32.0
django-prometheus==2.3.1

# Environment & Configuration
python-decouple==3.8
python-dotenv==1.0.0

# Email
django-anymail==10.0

# Caching
django-redis==5.3.0

# Admin interface enhancements
django-admin-interface==0.26.0
django-colorfield==0.9.0

# Forms and validation
django-crispy-forms==2.0
crispy-bootstrap5==0.7

# Timezone handling
pytz==2023.3
django-timezone-field==6.1.0

# UUID field support
django-uuid-upload-path==1.0.0

# Soft delete functionality
django-softdelete==0.11.5

# Audit logging
django-auditlog==2.3.0

# Rate limiting
django-ratelimit==4.1.0

# Health checks
django-health-check==3.17.0

# Feature flags
django-waffle==4.0.0